Test fixes

+++ train-interfaces/src/main/java/hu/bme/mit/train/interfaces/TrainController.java
@@ -4,9 +4,9 @@ public interface TrainController {
 
 	void followSpeed();
 
-	int getReferenceSpeed();
+	double getReferenceSpeed();
 
-	void setSpeedLimit(int speedLimit);
+	void setSpeedLimit(double speedLimit);
 
 	void setJoystickPosition(int joystickPosition);
